user の redux 関連モジュールを typescript-fsa で書き直し
